clang-tidy

+++ src/gpl/src/nesterovBase.cpp
@@ -772,7 +772,7 @@ void BinGrid::initBins()
 
   log_->info(GPL, 29, "BinSize: {} {}", binSizeX_, binSizeY_);
 
-  size_t bin_dimensions = binCntX_ * binCntY_;
+  size_t bin_dimensions = static_cast<size_t>(binCntX_ * binCntY_);
   // initialize bins_ vector
   bins_.reserve(bin_dimensions);
   int x = lx_, y = ly_;
